requirements.txt
fastapi[standard]
pymongo[srv]
pyjwt[crypto]
python-dotenv
dotenv
uvicorn
bcrypt
black
pydantic_settings
mypy
faker